improved features
- 불필요한 검색 페이지 코드 삭제, 기존에 검색한 항목 유지를 위한 템플릿 코드 추가.
- 기존에는 댓글, 게시글 갯수를 파악하기 위해 엔티티들을 컬렉션으로 읽어오고 그 size를 측정.
- Spring Data JPA의 count 기능을 사용하여 더 직관적으로 변경.
- 게시글 서비스에서 댓글 서비스 분리
